requirements.txt
anthropic==0.8.1
pdf2image==1.17.0
Pillow==10.2.0
python-dotenv==1.0.1
